Introduce a "Veteran Bodyguard" role.
I've received feedback that not giving any information on whether a Bodyguard's protection was effective (or needed at all) is "unrealistic".  I think this is a legitimate complaint, but giving out such information would make the Bodyguard an even stronger player, so better split it out to a separate role?